\name{DataSimilarity-package}
\alias{DataSimilarity-package}
\encoding{UTF-8}
\docType{package}
\title{
\packageTitle{DataSimilarity}
}
\description{
\packageDescription{DataSimilarity}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DataSimilarity}
\packageIndices{DataSimilarity}
The package provides various methods for comparing two or more datasets or their underlying distributions. Often, a permutation or asymptotic test for the null hypothesis of equal distributions \eqn{H_0: F_1 = F_2} or \eqn{H_0: F_1 = \dots = F_k} is performed.  
}
\author{
\packageAuthor{DataSimilarity}

Maintainer: \packageMaintainer{DataSimilarity}
}
\references{
Stolte, M., Kappenberg, F., Rahnenführer, J., Bommert, A. (2024). Methods for quantifying dataset similarity: a review, taxonomy and comparison. Statist. Surv. 18, 163 - 298. \doi{10.1214/24-SS149} 

Stolte, M., Kappenberg, F., Rahnenführer, J. & Bommert, A.  (2024). A Comparison of Methods for Quantifying Dataset Similarity. \url{https://shiny.statistik.tu-dortmund.de/data-similarity/}
}
\keyword{ package }
\keyword{ htest }
\keyword{ multivariate }
\keyword{ nonparametric }
\concept{ dataset similarity }

